+++ src/traits/AdminDatabaseTableTrait.php
@@ -7,6 +7,10 @@ use luya\admin\models\Lang;
 use luya\admin\models\NgrestLog;
 use luya\admin\models\QueueLog;
 use luya\admin\models\QueueLogError;
+use luya\admin\models\StorageFile;
+use luya\admin\models\StorageFilter;
+use luya\admin\models\StorageFilterChain;
+use luya\admin\models\StorageImage;
 use luya\admin\models\Tag;
 use luya\admin\models\TagRelation;
 use luya\admin\models\User;
@@ -472,6 +476,90 @@ trait AdminDatabaseTableTrait
         return new ActiveRecordFixture([
             'modelClass' => QueueLogError::class,
             'fixtureData' => $fixtureData,
+        ]);
+    }
+
+    /**
+     * Storage File
+     *
+     * @param array $fixtureData
+     * @return ActiveRecordFixture
+     * @since 1.2.0
+     */
+    public function createAdminStorageFileFixture(array $fixtureData = [])
+    {
+        return new ActiveRecordFixture([
+            'modelClass' => StorageFile::class,
+            'fixtureData' => $fixtureData,
+        ]);
+    }
+
+    /**
+     * Storage Image
+     *
+     * @param array $fixtureData
+     * @return ActiveRecordFixture
+     * @since 1.2.0
+     */
+    public function createAdminStorageImageFixture(array $fixtureData = [])
+    {
+        return new ActiveRecordFixture([
+            'modelClass' => StorageImage::class,
+            'fixtureData' => $fixtureData,
+        ]);
+    }
+
+    /**
+     * Storage Filter
+     *
+     * @param array $fixtureData
+     * @return ActiveRecordFixture
+     * @since 1.2.0
+     */
+    public function createAdminStorageFilterFixture(array $fixtureData = [])
+    {
+        return new ActiveRecordFixture([
+            'modelClass' => StorageFilter::class,
+            'fixtureData' => $fixtureData,
+        ]);
+    }
+
+    /**
+     * Storage Filter Chain
+     *
+     * @param array $fixtureData
+     * @return ActiveRecordFixture
+     * @since 1.2.0
+     */
+    public function createAdminStorageFilterChainFixture(array $fixtureData = [])
+    {
+        return new ActiveRecordFixture([
+            'modelClass' => StorageFilterChain::class,
+            'fixtureData' => $fixtureData,
+        ]);
+    }
+
+    
+
+
+    /**
+     * Create the Admin Queue Table
+     *
+     * @since 1.1.1
+     */
+    public function createAdminQueueTable()
+    {
+        return $this->createTableIfNotExists('admin_queue', [
+            'id' => 'INT(11) PRIMARY KEY',
+            'channel' => 'text',
+            'job' => 'blog',
+            'pushed_at' => 'int(11)',
+            'ttr' => 'int(11)',
+            'delay' => 'int(11)',
+            'priority' => 'int(11)',
+            'reserved_at' => 'int(11)',
+            'attempt' => 'int(11)',
+            'done_at' => 'int(11)',
         ]);
     }
 }